+++ lumicks/pylake/detail/imaging_mixins.py
@@ -239,10 +239,14 @@ class VideoExport:
                 )
                 return plt.gca().get_children()
 
-        line_ani = animation.FuncAnimation(
-            fig, plot, stop_frame - start_frame, interval=1, blit=True
+        fig = plt.figure()
+        fig.patch.set_alpha(1.0)
+        # Don't store the FuncAnimation in a variable as this leads to mpl attempting to remove
+        # a callback that doesn't exist on plt.close(fig) when using the jupyter notebook backend.
+        animation.FuncAnimation(fig, plot, stop_frame - start_frame, interval=1, blit=True).save(
+            file_name,
+            writer=writer,
         )
-        line_ani.save(file_name, writer=writer)
         plt.close(fig)