+++ lumicks/pylake/tests/test_file_download.py
@@ -21,8 +21,8 @@ def test_download_record_metadata():
     record = download_record_metadata("4280789")  # Older version of Pylake
 
     # Verify that the fields we rely on stay the same
-    assert record["files"][0]["checksum"] == "1a401193ab22f0983f87855e2581075b"
-    assert record["files"][0]["filename"] == "lumicks/pylake-v0.7.1.zip"
+    assert record["files"][0]["checksum"] == "md5:1a401193ab22f0983f87855e2581075b"
+    assert record["files"][0]["key"] == "lumicks/pylake-v0.7.1.zip"
     assert record["files"][0]["links"]["self"].startswith("https://zenodo.org/")  # Link may change
 
 
@@ -33,20 +33,24 @@ def test_non_zenodo_doi():
 
 
 @pytest.mark.preflight
-def test_download_from_doi(tmpdir_factory):
+def test_download_from_doi(tmpdir_factory, capsys):
     tmpdir = tmpdir_factory.mktemp("download_testing")
     record = download_record_metadata("4247279")
+
     files = download_from_doi("10.5281/zenodo.4247279", tmpdir, show_progress=False)
 
+    captured = capsys.readouterr()
+    assert not captured.out
+
     # Validate checksum
-    assert verify_hash(files[0], record["files"][0]["checksum"])
+    assert verify_hash(files[0], *record["files"][0]["checksum"].split(":"))
 
     # Add a random character such that the checksum fails
     with open(files[0], "ab") as f:
         f.write(b"\x21")
 
     # Validate that the hash is no longer correct
-    assert not verify_hash(files[0], record["files"][0]["checksum"])
+    assert not verify_hash(files[0], *record["files"][0]["checksum"].split(":"))
 
     with pytest.raises(
         RuntimeError,
@@ -57,5 +61,14 @@ def test_download_from_doi(tmpdir_factory):
 
     download_from_doi("10.5281/zenodo.4247279", tmpdir, force_download=True, show_progress=False)
 
+    captured = capsys.readouterr()
+    assert not captured.out
+
     # Validate checksum after forced re-download (should be OK again)
-    assert verify_hash(files[0], record["files"][0]["checksum"])
+    assert verify_hash(files[0], *record["files"][0]["checksum"].split(":"))
+
+    download_from_doi("10.5281/zenodo.4247279", tmpdir)
+
+    # Validate that we report that it was already downloaded
+    captured = capsys.readouterr()
+    assert r"Already downloaded cas9_kymo_compressed.h5" in captured.out