+++ changelog.md
@@ -12,6 +12,7 @@
 
 * Kymographs consisting of a single scan line now return a valid `line_time_seconds`. This allows certain downstream functionality, such as `Kymo.plot()`.
 * Issue a more descriptive error when attempting to compute a diffusion constant of a track with no points.
+* Pylake can now handle kymographs that were erroneously stored in the `Scan` field. Kymographs with a pre-specified number of lines to record were incorrectly being marked on the timeline and exported as 'Scan' instead of 'Kymograph' in versions of Bluelake prior to `2.5.0`.
 
 ## v1.2.1 | 2023-10-17